Allocator: refactor virtual memory management around new FSystemPageAllocator
many low-level systems handle a custom pool of meta data descriptor using
low-level FVirtualMemory API.
but allocation granularity was an issue: every custom needed to allocate
64k, or now 256k on Linux, every time. the allocator can is pool of system
pages that can distribute 4k pages, *on every plaform* (versus varying granularity)
some code needed to be refactored to handle the new API, but it should be cleaner now.
notably for TCompressedRadixTrie<>, which is now able to release allocated memory (yeah no more leaks)
some trickery add to be made with memory tracking to keep interesting stats, buts
the damage is controlled.

+++ Source/Runtime/Core/Private/Allocator/MallocBinned2.cpp
@@ -636,12 +636,12 @@ FBinnedSmallTable::FBinnedSmallTable() NOEXCEPT {
     ONLY_IF_MEMORYDOMAINS(BinnedTrackingDataStart_());
 #endif
 
-    const u32 osPageSize = PPE_MALLOCBINNED2_OS_PAGESIZE;
-    const u32 osGranularity = PPE_MALLOCBINNED2_OS_GRANULARITY;
-    const u32 maxChunkSize = PPE_MALLOCBINNED2_CHUNK_MAXSIZE;
-    const u32 chunkOverheadSize = PPE_MALLOCBINNED2_CHUNK_OVERHEADSIZE;
-    const u32 minBlocksPerChunk = PPE_MALLOCBINNED2_CHUNK_MINBLOCKS;
-    const u32 maxBlocksPerChunk = PPE_MALLOCBINNED2_CHUNK_MAXBLOCKS;
+    const u32 osPageSize = checked_cast<u32>(PPE_MALLOCBINNED2_OS_PAGESIZE);
+    const u32 osGranularity = checked_cast<u32>(PPE_MALLOCBINNED2_OS_GRANULARITY);
+    const u32 maxChunkSize = checked_cast<u32>(PPE_MALLOCBINNED2_CHUNK_MAXSIZE);
+    const u32 chunkOverheadSize = checked_cast<u32>(PPE_MALLOCBINNED2_CHUNK_OVERHEADSIZE);
+    const u32 minBlocksPerChunk = checked_cast<u32>(PPE_MALLOCBINNED2_CHUNK_MINBLOCKS);
+    const u32 maxBlocksPerChunk = checked_cast<u32>(PPE_MALLOCBINNED2_CHUNK_MAXBLOCKS);
 
     for (u32 i = 0; i < PPE_MALLOCBINNED2_SMALLPOOL_COUNT; ++i) {
         FSmallPoolInfo& pool = SmallPools[i];
